#6: Add heart rate zones to the charts
The zones are displayed as colors in the background, and are hidden if the heart rate chart itself is hidden.

Pace zones cannot be shown using this method as the background is shared. Will have to think of a better way to handle them (show if pace is shown and HR isn't?).

Still no pause markers shown for the distance-based chart.